+++ src/linter/config-presets/aglint-recommended.ts
@@ -13,6 +13,7 @@ const config: LinterConfig = {
         'if-closed': 'error',
         'inconsistent-hint-platforms': 'error',
         'invalid-domain-list': 'error',
+        'invalid-modifiers': 'error',
         'unknown-hints-and-platforms': 'error',
         'unknown-preprocessor-directives': 'error',
     },