+++ src/linter/rules/index.ts
@@ -5,6 +5,7 @@ import { UnknownHintsAndPlatforms } from './unknown-hints-and-platforms';
 import { DuplicatedModifiers } from './duplicated-modifiers';
 import { IfClosed } from './if-closed';
 import { InvalidDomainList } from './invalid-domain-list';
+import { InvalidModifiers } from './invalid-modifiers';
 import { InconsistentHintPlatforms } from './inconsistent-hint-platforms';
 import { SingleSelector } from './single-selector';
 import { UnknownPreProcessorDirectives } from './unknown-preprocessor-directives';
@@ -18,5 +19,6 @@ export const defaultLinterRules = new Map<string, AnyLinterRule>([
     ['duplicated-hints', DuplicatedHints],
     ['unknown-hints-and-platforms', UnknownHintsAndPlatforms],
     ['invalid-domain-list', InvalidDomainList],
+    ['invalid-modifiers', InvalidModifiers],
     ['inconsistent-hint-platforms', InconsistentHintPlatforms],
 ]);